Move generate puzzle logic to ISA
