+++ src/isa/parallel/execute_state.rs
@@ -114,7 +114,7 @@ impl ExecuteState {
             self.level_index,
             statistics,
             best,
-            self.editor,
+            Box::new(self.editor),
           )));
         }
