+++ src/level.rs
@@ -1,4 +1,3 @@
-use rlua::prelude::*;
 use serde::Deserialize;
 use std::error::Error;
 use std::fmt;
@@ -10,7 +9,6 @@ use std::path::Path;
 use uuid::Uuid;
 
 use crate::global_state::GlobalState;
-use crate::puzzle::{Puzzle, TestCaseSet};
 
 /// Stores all details about the levels
 #[derive(Debug, Clone, Default, Deserialize)]
@@ -161,47 +159,6 @@ impl Level {
     } else {
       format!("Level {} - {}", level_index, self.name())
     }
-  }
-
-  ///
-  /// Load and run the Lua code to generate the puzzles
-  ///
-  pub fn generate_test_cases(&self, seed: u32, n: usize) -> Result<TestCaseSet, Box<dyn Error>> {
-    // Try to load the Lua code file into memory
-    let lua_code = fs::read_to_string(format!("levels/{}", self.lua_file))?;
-
-    // Generate and run the code within the Lua context
-    let test_cases = Lua::new().context::<_, LuaResult<TestCaseSet>>(|ctx| {
-      let globals = ctx.globals();
-
-      // Add the levels folder to the path
-      ctx
-        .load(&r#"package.path = "./levels/?.lua;" .. package.path"#)
-        .exec()?;
-
-      // Seed the random number generator
-      globals
-        .get::<_, LuaTable>("math")?
-        .get::<_, LuaFunction>("randomseed")?
-        .call::<_, ()>(seed)?;
-
-      // Load the script code
-      //  This should define a global function named "generateTestCase"
-      ctx.load(&lua_code).exec()?;
-
-      // Generate the test cases one-by-one
-      let generate_test_case: LuaFunction = globals.get("generateTestCase")?;
-      let test_cases = (0..n)
-        .map(|_| {
-          let (inputs, outputs): (Vec<i16>, Vec<i16>) = generate_test_case.call(())?;
-          Puzzle::new(inputs, outputs).map_err(LuaError::RuntimeError)
-        })
-        .collect::<Result<_, _>>()?;
-
-      Ok(test_cases)
-    })?;
-
-    Ok(test_cases)
   }
 }
