+++ src/state/mod.rs
@@ -1,8 +1,6 @@
 use crossterm::{cursor, QueueableCommand};
 use std::io::{self, Write};
 
-mod editor_state;
-mod execute_state;
 mod level_select_state;
 mod show_help_state;
 mod state;
